feat(chain-of-responsibility): Add new patter
The Chain of responsibility standard falls as a specific list, and is also the following in the
scenario that must be validated, otherwise it is not a condition. In such cases, the Chain of
Responsibility enables us to separate responsibilities from small and lean classes, and provides a
flexible and uncoupled way to bring these behaviors back together.